5. Módulo de Banca/requirements.txt
spyne
mysql-connector-python